Separated sub-types from HNode into their own files for easier maintenance
